+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 0.9 (unreleased)
 ----------------
 
+- Remove question for locales.
+  [timo]
+
 - Remove questions for version and license.
   [timo]